Fix cost tracking precision using Decimal instead of integer cents
Change cost_cents from integer to decimal (precision: 12, scale: 4) to
properly track fractional cents for low-cost AI operations. Previously,
small costs like 0.06 cents would round to 0, causing Monthly Cost to
display as $0.00.

- Update migrations, schemas, and queries to use Decimal
- Fix arithmetic in Usage module with Decimal.add/mult/compare
- Add to_number helper in LiveView for chart calculations
- Update tests to use Decimal.equal? for comparisons
- Improve export/import scripts with better error handling

+++ lib/diagram_forge/usage/alert.ex
@@ -13,7 +13,7 @@ defmodule DiagramForge.Usage.Alert do
   schema "usage_alerts" do
     field :period_start, :date
     field :period_end, :date
-    field :amount_cents, :integer
+    field :amount_cents, :decimal
     field :email_sent_at, :utc_datetime
     field :acknowledged_at, :utc_datetime
